feat: spit SetLogs model from sets file
